requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.15
django-filter==21.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
model-bakery==1.7.0
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.3
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zipp==3.6.0
